Initial (i.e. largely unusable) support for adding/deleting rows.
This manipulation occurs in reaction to the data fed from the server.
Torrents in the incoming data that are not present in the page are
inserted at the end of the table.  Torrents that are not present in
the data have their rows deleted.  Existing torrents are updated.

Or at least, that's the plan: this commit is probably hosed.