Initial (i.e. largely unusable) support for adding/deleting rows.
This manipulation occurs in reaction to the data fed from the server.
Torrents in the incoming data that are not present in the page are
inserted at the end of the table.  Torrents that are not present in
the data have their rows deleted.  Existing torrents are updated.

Or at least, that's the plan: this commit is probably hosed.

A bit more advanced on the table manipulation, but not ready yet.

Use internally generated IDs for page transitions and actions.

The IDs generated by ctorrent aren't particularly reliable for HTTP
transit: many awkward characters (single quotes, backticks, etc.) are
permissiable, which makes assembling a parsable AJAX string a pain in
the arse.

Now, once a new torrent arrives, it is assigned an internally 'generated'
ID, currently an ever-incrementing integer.  Yes, this will wrap at an
obscenely large value.  I don't care.

Misc bugfixes.

More bugfixes.

This time around the AJAX-caused creation of new rows.