Lose the 'dump' debugging statements.
Get rid of all those 'class="numeric"' table datas.

They really mean 'centered' in actual usage, and are the norm, not
the exception.  Only the titles are left-aligned, so single those
out as the special case, and make TDs centered by default.

Mostly a tidy-up, but with a fix for table row colours.

Sorting the rows meant that the row colours could get screwed up,
with the alternating pattern being lost.  This fixes that by
adjusting the row class as the rows are re-inserted into the
correct places.

Highlight completed torrents by making them bold.

Style adjustment for advanced limits page.