Improved handling of message functions.
function_to_handle_request no longer returns null_message_function
invocations if no message is appropriate (which happens for a great
many messages).  This *greatly* improves performance, as many
redundant function objects are not created and invoked.

Instead, an 'undef' value indicates that there is no suitable
response.  An exception indicates that the incoming message could
not be handled.