Fix an unbelievably stupid mistake with sending messages.
Keys of hashes are not object references: they are strings.
It is therefore not valid to send a message to a key retrieved
by finding that its value matches some criteria.

Removing the null_writer_function exposes this, as it was the only
periodic function being sent to a real socket: this had the effect
of draining the message queue that had been filled using the
string value of a socket object (AARGGH!!).  Without that function
being sent, $writers was empty (can't add a string!), so the messages
just built up.

On the plus side, this removes the need for null_writer_function,
which in turn seriously reduces the number of empty valued,
anonymous subs being created, making the whole thing much, MUCH
lighter weight.  Also, messages sent to ctorrent instances that
weren't saying anything would get stuck since the null_writer_function
wasn't unclogging the message queue: this doesn't happen anymore.