Fix long-standing annoyance with percent_complete.
sprintf format conversion problems with the JSON string.  Adding
a percentage symbol there seems weird: the percent_complete value
should just be an integer.  This means that there are now two places
in the JavaScript (once for updates, and once for inserts) where
a '%' symbol is appended to the cell value.