torrent-status API function.
Cleaned up 'apicall' to remove the second (name) parameter, as
it was the same as the hash key and the duplication was annoying
me: the reporting string now comes from the caller, which is
enough.

torrent-status returns a JSON formatted string containing the
status of either all registered torrents, or just the ones with
ids given in the torrents HTTP parameter (as with other API
functions).  The JSON structure returned is a list of hashes,
with one hash per torrent.  This is the same structure that feeds
the AJAX calls that update the index page.

This required some refactoring of the JSON routines, but it
works with some preliminary testing.