Better 'quit' API call handling.  Expanded API
The previous 'quit' behaviour on the API returned nothing to the
client if the quit was unsuccessful because no torrent ids were
supplied.  The new call sends a JSON hash back to the client,
with the key 'error', and an error string as the value.

Also greatly expanded the API to cover most of the available API
functions.  I don't really want to make the 'exit' call available
for casual use (at least not until I get an authentication scheme
sorted).  The 'count' call is almost useless, as the 'torrents'
query has a 'count' field in the returned hash.  Not decided if
I want to list available API functions via the 'api' API call
yet.