Added 'set-config' API method.
This involves changes to the server to be able to process this
effectively.  At the moment, only one torrent at a time can
be processed, as identifying which parameters belong to which
torrent involves some kludgy string-chopping.

Since the contents of config are now available externally, I've
added a validator function to the 'validator' key for each config
option: when supplied with a proposed value, it will respond to
indicate whether or not the value is valid.  This validator
function is automatically called by 'torrent_set_config' prior
to applying new values.