Bring responses to unknown API methods into line
Just as API methods have been pulled into the standard plumbing,
this commit does the same for responses to unknown API calls.

This completely removes the need for send_http_message, so I've
deleted this function.  prepare_http_message is still useful,
although perhaps overkill.  It's better just to leave it as-is,
though, as it's working fine.