Improved creation of config validation functions.
And configuration setting in general.  Previously, I'd been
overwriting the previous config element associated with the
torrent, taking pains to carefully delete the validator
function and then re-creating the same thing.  Now, the validator
is created ONCE for a particular option.  It's still carefully
detached when the client socket is de-registered.