More consistent responses to API calls.
Blocks given to 'apicall' now have a simple contract:

  * return a data structure to encode as a response on success
  * die with a message on failure

Combined with each apicall function being invoked with its name,
this allows for 'action' and 'status' messages to be consistently
tagged on to every response, giving the client a better idea
of whether a given response is good or bad.  Currently, the
status message is limited to 'success' or 'failure'.