Cleaned up responding to requests for HTML vs JSON
Previously handled by passing along a reference to a scalar
that indicated whether or not the page content should be wrapped
in a normal HTML page, or just sent as-is.

The new code only does the wrapping--via html_page--if asked to,
and provides a more flexible way to generate content in response
to requests.