Better pretty_print_json_string
Is now pretty_print_json_string_ref, and uses IO::Scalar to
operate on a given string reference rather than IO::String working
on a (possibly very large) string.

The generation of the /debug page is horribly slow and expensive,
largely because of the loop in that method that's executed for
*every* character in the text (I suspect).  Need to do some
profiling to see where the time is going, and try a few possible
optmisiations out.