Hide the favicon in a sub attached to %ctcs
$ctcs{favicon} is a CODE ref that evaluates to the favicon itself.
This keeps the state for the server split between %clients and
%ctcs, removing the now redundant $favicon global.  The CODE ref
is good because it hides the favicon data in its closure, which
doesn't print out in full in the /debug page.  Simply attaching
the favicon to the %ctcs hash would result in a ugly splurge of
semi-printable data on that page.