Account for EINTR.  IP default for address_of
EINTR can pop up with system calls, but is rarely fatal.
Interrupting read or write calls (sysread and syswrite) is unlikely,
but possible.  Interruption of the select call is far more likely.

In any of thee cases, I handle EINTR by logging it, but otherwise
ignoring it: a problem in dealing with a single client should not
take the server down.

Additionally, address_of now defaults to the dot-notation IP
address of the peer in question if the DNS lookup fails.  I should
really have done that some time ago.