dynamic dispatch with request handlers hash
This is an experimental hack, putting handlers for individual
web or ctorrent requests in a dispatch table attached to %ctcs.

I'm not 100% convinced about this yet, mostly because of the hack
in client_request_handler that puts three variables into the
package namespace (although it does make the individual handlers
nicer to look at).

I may convert these handlers to use @_ in the same way that
apicalls do at some point in the future.